chat_with_data/requirements.txt
streamlit==1.28.0
openai==1.3.0
sqlalchemy==2.0.23
pandas==2.1.3
plotly==5.18.0
python-dotenv==1.0.0
PyGithub==2.1.1

